django-api/requirements.txt
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
